connections: add support for arbitrary code execution
In a connection editor, users can now select "<statement(s)>" in place of a method and enter a multiline code expression into the pane. The expression has access to "from" and "to" and is rendered amongst other connections to the target. The code expression is rendered in the connections list. Users can later change a connection between expression and invocation mode. Undo/redo is supported as well.

Breaking API changes:
- Moved Utils.print_connection() to Connection.print()

Co-authored-by: Philipp Keese <philipp.keese@student.hpi.de>