Add multiline / side-effect Expressions
This change requires every connection script (except Expressions) to include a return statement by themselves in order to have separate code executed prior to that return statement. We fixed the car example for this, however you might need to add a return statement to connection arguments of your existing prototypes.